+++ apps/src/javalab/neighborhood/neighborhood-visualization-column.module.scss
@@ -1,0 +1,7 @@
+.collapsed {
+  opacity: 0;
+}
+
+.expanded {
+  opacity: 1;
+}
