+++ apps/src/miniApps/neighborhood/neighborhood.module.scss
@@ -1,0 +1,5 @@
+.neighborhoodPreviewBackground {
+  background-repeat: no-repeat;
+  background-size: cover;
+  background-position: top;
+}
